add get-by-email
