+++ auth-db/queries/email-auth/index.ts
@@ -1,1 +1,2 @@
 export * from "./create.ts";
+export * from "./get-by-email.ts";
